+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Homepage lookup statistics gain a range selector (24 hours default / 7 days / 30 days / all time) with per-range totals and top countries.
+
 ## [1.8.0] - 2026-08-20
 
 ### Changed